+++ src/main/java/za/co/discovery/application/dto/TransactionalAccountBalancesDTO.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class TransactionalAccountBalancesDTO {
 	 private  String AccountNumber;
 	 private String AccountType;
-	 private BigDecimal AccountBalance;
+	 private Float AccountBalance;
 }
